requirements.txt
alembic==1.7.1
docutils==0.17.1
Flask==2.0.1
flask-cors==3.0.10
flask-login==0.5.0
Flask-WTF==0.15.1
googleapis-common-protos==1.53.0
importlib_resources==1.4.0
mypy-protobuf==2.9
protobuf==3.17.3
pyln-client==0.10.1
PySocks==1.7.1
python-bitcoinlib==0.11.0
pyzmq==22.3.0
requests[socks]==2.26.0
SQLAlchemy==1.4.25
squeaklib==0.13.1
typed-config==0.2.5
Werkzeug==2.0.3
